+++ headers/meta.hpp/meta_detail/type_traits/enum_traits.hpp
@@ -12,7 +12,7 @@
 
 namespace meta_hpp::detail
 {
-    enum class enum_flags : std::uint32_t {
+    enum class enum_flags : std::uint8_t {
         is_scoped = 1 << 0,
     };
